# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        maxLength: 72
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      is_admin_group:
        type: boolean
      is_protected:
        type: boolean
      name:
        maxLength: 50
        minLength: 3
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
      role_id:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      last_failed_login:
        type: string
      last_login_at:
        type: string
      password_changed_at:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: WattWatch API server.
  title: WattWatch API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: List users with optional filtering. Requires admin privileges.
      parameters:
      - description: Search by username or email
        in: query
        name: search
        type: string
      - description: Filter by role ID
        in: query
        name: role_id
        type: string
      - description: Field to order by (username, email, created_at)
        in: query
        name: order_by
        type: string
      - description: Order descending
        in: query
        name: order_desc
        type: boolean
      - description: 'Limit results (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Offset results (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - admin only
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users (Admin only)
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user. Users can only delete their own account unless they
        are an admin.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - can only delete own account unless admin
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user by their ID (requires auth, users can only access their
        own profile unless admin)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - can only access own profile unless admin
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's details
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: User details to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Change a user's password (users can only change their own password)
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Password change details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid request or password requirements not met
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Permission denied - can only change own password
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Password was recently used
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"